# docs/swagger.yaml
basePath: /v1/api
definitions:
...
  go-app_internal_dto.CreateMessageDto:
    properties:
      channel_id:
        type: string
      content:
        type: string
      msg_type:
        type: string
      replied_to_msg_id:
        type: string
    required:
    - channel_id
    - content
    - msg_type
    type: object
...
  go-app_internal_dto.CreateTokenDto:
    properties:
      role:
        type: string
      userEmail:
        type: string
      userId:
        type: string
    type: object
...
  go-app_internal_dto.GeneratePresignedURLReq:
    properties:
      content_type:
        type: string
      folder:
        type: string
      object_name:
        type: string
    required:
    - content_type
    - object_name
    type: object
...
  go-app_internal_dto.UpdateMessageDto:
    properties:
      content:
        type: string
      status:
        type: string
    type: object
  go-app_internal_dto.UpdateUserDto:
    properties:
      avatar_url:
        type: string
      is_active:
        type: boolean
      password:
        type: string
      user_name:
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: This is a sample server API for Go App.
  title: Go App API
  version: "1.0"
paths:
...
  /channel/{channel-id}/history:
    delete:
      description: Delete chat history in a channel up to a sequence
      parameters:
      - description: Channel ID
        in: path
        name: channel-id
        required: true
        type: string
      - description: Up To Sequence
        in: query
        name: up_to_seq
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete chat history
      tags:
      - message
...
  /channel/{channel-id}/messages:
    get:
      description: Get messages in a channel with pagination
      parameters:
      - description: Channel ID
        in: path
        name: channel-id
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: string
      - description: Before Sequence
        in: query
        name: before_seq
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get messages by channel
      tags:
      - message
...
  /message:
    post:
      consumes:
      - application/json
      description: Create a new message in a channel
      parameters:
      - description: Message Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.CreateMessageDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Create message
      tags:
      - message
  /message/{message-id}:
    get:
      description: Get a single message by ID
      parameters:
      - description: Message ID
        in: path
        name: message-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get message by ID
      tags:
      - message
    put:
      consumes:
      - application/json
      description: Update message content
      parameters:
      - description: Message ID
        in: path
        name: message-id
        required: true
        type: string
      - description: Update Message Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.UpdateMessageDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update message
      tags:
      - message
  /message/{message-id}/hide:
    post:
      description: Hide a message for current user
      parameters:
      - description: Message ID
        in: path
        name: message-id
        required: true
        type: string
      - description: Channel ID
        in: query
        name: channel_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Hide message
      tags:
      - message
  /message/{message-id}/recall:
    delete:
      description: Recall a message
      parameters:
      - description: Message ID
        in: path
        name: message-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Recall message
      tags:
      - message
  /rf/refresh:
    post:
      consumes:
      - application/json
      description: Create a new refresh token
      parameters:
      - description: Token Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.CreateTokenDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Create refresh token
      tags:
      - refresh-token
...
  /upload/presigned:
    post:
      consumes:
      - application/json
      description: Generate a presigned URL for file upload
      parameters:
      - description: Presigned URL Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.GeneratePresignedURLReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Generate presigned URL
      tags:
      - upload
  /user:
    put:
      consumes:
      - application/json
      description: Update current user information
      parameters:
      - description: Update User Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.UpdateUserDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update user
      tags:
      - user
  /user/{user-id}:
    delete:
      description: Delete user by ID
      parameters:
      - description: User ID
        in: path
        name: user-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete user
      tags:
      - user
    get:
      description: Get user information by their ID
      parameters:
      - description: User ID
        in: path
        name: user-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get user by ID
      tags:
      - user
...
  /user/search:
    get:
      description: Search users by name
      parameters:
      - description: Search Keyword
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Search users
      tags:
      - user
...
swagger: "2.0"